fix: code quality, test robustness, and docs accuracy across monorepo
nars-api: add number-entrances service (NumberEntrancesService) wired via
ServiceRegistrationExtensions; drop dead IFeatureService dependency in
DraftFeaturesController; tidy LogSanitizer and feature queries.

nars-tests: extract shared ServiceTestBase to remove 9-way fixture
boilerplate; await GetEntranceNumberAsync instead of sync-over-async;
exclude PostGIS system table spatial_ref_sys from CleanTablesAsync;
align tests with refactored services.

nars-infra: document-only k8s fixes (postgis init-chown resources,
ingress hostless-rule trim, network-policy egress scope comment,
backup-cronjob lifetime comment, secret key doc, roads service port name)
and add roads/secrets.yaml reference template (no stringData).

nars-roads: config/model quality fixes.

nars-web: fix house-numbering race condition (await road-side before
assigning), static map-init import, i18n keys.

make: single-request smoke check in smoke.mk; add docs-tex-pdf target
and correct schema-table drift note in make/docs.mk.

docs: expand database schema table from 15 to 18 tables (error_logs,
inspections, ai_draft_features) and fix dairas/communes FK + index rows
to match the migrations; 378 module units unaffected.

+++ nars-api/DTOs/FeatureDtos.cs
@@ -71,7 +71,7 @@ public record CreateResponse(
 
 public record LoadFeaturesResponse<T>(
     IReadOnlyList<T> Features,
-    int Count,
+    long Count,
     int Skip,
     int Take
 );
